Add script to build icons
Add a script that is used to composite the icon of the action with the
background of the button (btn-background.png). It makes use of the "composite"
tool, which in ubuntu is provided by the imagemagick package.

Also, add two sample icons for GPS enable/disable (better ones coming soon!)